Modifications to allow override of value getters
Modified forms.js to allow overriding of form element getters. This was
to resolve issue where overriding the control rendering closure for a
particular element id to a control of a completely different type (think
checkbox to yes/no drop down) would be unable to get the value of the
overridden element.