Fix to deal with 'radio textboxes'
Sometimes pdfs are created (apparently by state gov'ts) with 'radio textboxes' which are standard text boxes (or drop downs etc) that have the grouping `.[\d] as part of full name. This is wrong and bad, and nearly invisible in most pdf renderers and editors, but here we are. This fix strips the grouping and sets a corrected id for these as well without it.

Otherwise data tends not to be properly set.