Fix to replicated form fields
Base library improperly dealt with 'duplicated' fields, aka a text field that was replicated on the form. Only the first is treated as valid per library, and getting data only pulls from first. Only the first element is editable, and others are locked. This was applied only to text elements. Getting into this on checkboxes, dropdowns and pushbuttons was decidedly weird and is best left to redesigning the original pdf document.